added delete logic

+++ src/main/java/com/sequenceiq/provisioning/service/SimpleInfraService.java
@@ -1,20 +1,24 @@
 package com.sequenceiq.provisioning.service;
 
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 import java.util.UnknownFormatConversionException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.sequenceiq.provisioning.controller.BadRequestException;
 import com.sequenceiq.provisioning.controller.NotFoundException;
 import com.sequenceiq.provisioning.controller.json.InfraJson;
 import com.sequenceiq.provisioning.converter.AwsInfraConverter;
 import com.sequenceiq.provisioning.converter.AzureInfraConverter;
 import com.sequenceiq.provisioning.domain.AwsInfra;
 import com.sequenceiq.provisioning.domain.AzureInfra;
+import com.sequenceiq.provisioning.domain.CloudInstance;
 import com.sequenceiq.provisioning.domain.Infra;
 import com.sequenceiq.provisioning.domain.User;
+import com.sequenceiq.provisioning.repository.CloudInstanceRepository;
 import com.sequenceiq.provisioning.repository.InfraRepository;
 
 @Service
@@ -28,6 +32,9 @@ public class SimpleInfraService implements InfraService {
 
     @Autowired
     private AzureInfraConverter azureInfraConverter;
+
+    @Autowired
+    private CloudInstanceRepository cloudInstanceRepository;
 
     @Override
     public Set<InfraJson> getAll(User user) {
@@ -78,7 +85,16 @@ public class SimpleInfraService implements InfraService {
         if (infra == null){
             throw new NotFoundException(String.format("Infrastructure '%s' not found.", id));
         }
-        infraRepository.delete(infra);
+        List<CloudInstance> allCloudForInfra = getAllCloudForInfra(id);
+        if (allCloudForInfra.size() == 0) {
+            infraRepository.delete(infra);
+        } else {
+            throw new BadRequestException(String.format("Infrastructure '%s' has some cloud dependency please remove clouds before the deletion.", id));
+        }
+    }
+
+    private List<CloudInstance> getAllCloudForInfra(Long id) {
+        return cloudInstanceRepository.findAllCloudForInfra(id);
     }
 
 }
